+++ packaging/scripts/validate-packaging-metadata.sh
@@ -87,8 +87,11 @@ expect_line packaging/homebrew/Formula/slskdn.rb '^class Slskdn < Formula$'
 
 expect_line .github/workflows/release-packages.yml 'slskdn-main-linux-glibc-x64\.zip'
 expect_literal .github/workflows/build-on-tag.yml 'cp packaging/linux/install-from-release.sh release/install-linux-release.sh'
-expect_literal .github/workflows/build-on-tag.yml 'sha256sum *.zip slskd.service slskd.yml slskd.sysusers install-linux-release.sh > SHA256SUMS.txt'
-expect_literal .github/workflows/build-on-tag.yml 'cp packaging/aur/slskd.service packaging/aur/slskd.yml packaging/aur/slskd.sysusers release/'
+expect_literal .github/workflows/build-on-tag.yml 'sha256sum *.zip slskd.service slskd.yml slskd.sysusers slskd.tmpfiles install-linux-release.sh > SHA256SUMS.txt'
+expect_literal .github/workflows/build-on-tag.yml 'cp packaging/aur/slskd.service packaging/aur/slskd.yml packaging/aur/slskd.sysusers packaging/aur/slskd.tmpfiles release/'
+expect_literal .github/workflows/build-on-tag.yml 'cp ../packaging/aur/slskd.tmpfiles slskd.tmpfiles'
+expect_literal .github/workflows/build-on-tag.yml 'cp packaging/aur/slskd.tmpfiles ~/rpmbuild/SOURCES/'
+expect_literal .github/workflows/build-on-tag.yml 'cp packaging/aur/slskd.tmpfiles slskdn-${VERSION}/usr/lib/tmpfiles.d/slskd.conf'
 
 expect_line .github/workflows/release-packages.yml '\$\{\{ steps\.version\.outputs\.tag \}\}-linux-x64\.zip'